pyproject.toml
[project]
name = "doc-chat-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "fastapi>=0.133.1",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.10",
    "langchain-qdrant>=1.1.0",
    "langchain-text-splitters>=1.1.1",
    "langchain[openai]>=1.2.10",
    "langgraph>=1.0.9",
    "pypdf>=6.7.3",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "qdrant-client>=1.17.0",
    "sentence-transformers>=5.2.3",
    "slowapi>=0.1.9",
    "uvicorn>=0.41.0",
]
